fix(api): move SectorColors, SubSectorColors to own file
To prevent importing app-theme.ts in backend, which calls Chakra
functions that fail when executing on the server
