+++ app/src/lib/theme/app-theme.ts
@@ -15,20 +15,7 @@ import {
 } from "./recipes";
 import { buttonRecipe } from "./recipes/button.recipe";
 import { separatorRecipe } from "./recipes/separator.recipe";
-
-export enum SectorColors {
-  I = "#5785F4",
-  II = "#DF2222",
-  III = "#F28C37",
-  IV = "#2DD05B",
-  V = "#C6C61D",
-}
-
-export enum SubSectorColors {
-  "V.1" = "#38A857",
-  "V.2" = "#0E5221",
-  "V.3" = "#149037",
-}
+import { SectorColors } from "./custom-colors";
 
 export const appTheme = createSystem(defaultConfig, {
   globalCss: {
