+++ app/src/lib/theme/custom-colors.ts
@@ -1,0 +1,13 @@
+export enum SectorColors {
+  I = "#5785F4",
+  II = "#DF2222",
+  III = "#F28C37",
+  IV = "#2DD05B",
+  V = "#C6C61D",
+}
+
+export enum SubSectorColors {
+  "V.1" = "#38A857",
+  "V.2" = "#0E5221",
+  "V.3" = "#149037",
+}
